+++ bindings/core/src/method/client.rs
@@ -7,7 +7,7 @@ use iota_sdk::client::mqtt::Topic;
 use iota_sdk::{
     client::{
         api::{
-            ClientBlockBuilderOptions as BuildBlockOptions, GetAddressesBuilderOptions as GenerateAddressesOptions,
+            ClientBlockBuilderOptions as BuildBlockOptions, GetAddressesOptions as GenerateAddressesOptions,
             PreparedTransactionDataDto,
         },
         node_api::indexer::query_parameters::QueryParameter,
@@ -99,7 +99,7 @@ pub enum ClientMethod {
     },
     /// Generate addresses.
     #[serde(rename_all = "camelCase")]
-    GenerateAddresses {
+    GenerateEd25519Addresses {
         /// Create secret manager from json
         #[derivative(Debug(format_with = "OmittedDebug::omitted_fmt"))]
         secret_manager: SecretManagerDto,