+++ sdk/src/wallet/bindings/nodejs/examples/2-generate-address.js
@@ -10,15 +10,17 @@ async function run() {
         const account = await manager.getAccount('0');
         console.log('Account:', account);
 
-        const address = await account.generateAddress();
+        const address = await account.generateEd25519Address();
         console.log('New address:', address);
 
         // It's also possible to generate multiple addresses
-        // const addresses = await account.generateAddresses(2);
+        // const addresses = await account.generateEd25519Addresses(2);
         // console.log('New addresses:', addresses);
 
         // Use the Faucet to send testnet tokens to your address:
-        console.log("Fill your address with the Faucet: https://faucet.testnet.shimmer.network")
+        console.log(
+            'Fill your address with the Faucet: https://faucet.testnet.shimmer.network',
+        );
     } catch (error) {
         console.log('Error: ', error);
     }