+++ sdk/src/wallet/message_interface/response.rs
@@ -92,8 +92,13 @@ pub enum Response {
     /// [`SubmitAndStoreTransaction`](crate::wallet::message_interface::AccountMethod::SubmitAndStoreTransaction)
     SignedTransactionData(SignedTransactionDataDto),
     /// GenerateAddress response.
-    /// Response for [`GenerateAddresses`](crate::wallet::message_interface::AccountMethod::GenerateAddresses)
-    GeneratedAddress(Vec<AccountAddress>),
+    /// Response for
+    /// [`GenerateEd25519Addresses`](crate::wallet::message_interface::AccountMethod::GenerateEd25519Addresses)
+    GeneratedEd25519Addresses(Vec<AccountAddress>),
+    /// GenerateAddress response.
+    /// Response for
+    /// [`GenerateEvmAddresses`](crate::wallet::message_interface::AccountMethod::GenerateEvmAddresses)
+    GeneratedEvmAddresses(Vec<String>),
     /// Response for
     /// [`GetBalance`](crate::wallet::message_interface::AccountMethod::GetBalance),
     /// [`SyncAccount`](crate::wallet::message_interface::AccountMethod::SyncAccount)
@@ -162,7 +167,7 @@ pub enum Response {
     /// Response for [`Bech32ToHex`](crate::wallet::message_interface::Message::Bech32ToHex)
     HexAddress(String),
     /// Response for [`HexToBech32`](crate::wallet::message_interface::Message::HexToBech32)
-    /// Response for [`GenerateAddress`](crate::wallet::message_interface::Message::GenerateAddress)
+    /// Response for [`GenerateEd25519Address`](crate::wallet::message_interface::Message::GenerateEd25519Address)
     Bech32Address(Bech32Address),
     /// Response for
     /// [`RequestFundsFromFaucet`](crate::wallet::message_interface::AccountMethod::RequestFundsFromFaucet)
@@ -208,7 +213,8 @@ impl Debug for Response {
             Self::SignedTransactionData(signed_transaction_data) => {
                 write!(f, "SignedTransactionData({signed_transaction_data:?})")
             }
-            Self::GeneratedAddress(addresses) => write!(f, "GeneratedAddress({addresses:?})"),
+            Self::GeneratedEd25519Addresses(addresses) => write!(f, "GeneratedEd25519Addresses({addresses:?})"),
+            Self::GeneratedEvmAddresses(addresses) => write!(f, "GeneratedEvmAddresses({addresses:?})"),
             Self::Balance(balance) => write!(f, "Balance({balance:?})"),
             Self::SentTransaction(transaction) => write!(f, "SentTransaction({transaction:?})"),
             Self::MintTokenTransaction(mint_transaction) => {